minjun 브랜치와 master 브랜치 병합 작업

+++ app/src/main/java/com/ssu/gardenmaker/ui/MainActivity.kt
@@ -1,5 +1,7 @@
 package com.ssu.gardenmaker.ui
 
+import android.animation.ObjectAnimator
+import android.app.AlertDialog
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.graphics.Rect
@@ -7,6 +9,7 @@ import android.os.Build.VERSION_CODES
 import android.os.Bundle
 import android.os.Handler
 import android.os.Message
+import android.util.DisplayMetrics
 import android.util.Log
 import android.view.MenuItem
 import android.view.MotionEvent
@@ -66,6 +69,12 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         if (ContextCompat.checkSelfPermission(this@MainActivity, android.Manifest.permission.ACTIVITY_RECOGNITION) == PackageManager.PERMISSION_DENIED) {
             var permissions = arrayOf(android.Manifest.permission.ACTIVITY_RECOGNITION)
             requestPermissions(arrayOf(android.Manifest.permission.ACTIVITY_RECOGNITION),0)
+        }
+
+        // 뷰 중앙으로 이동
+        binding.mainLayout.scrollViewMain.post{
+            val mid = binding.mainLayout.scrollViewMain.width / ((this.resources.displayMetrics.densityDpi.toFloat()) / DisplayMetrics.DENSITY_DEFAULT)
+            binding.mainLayout.scrollViewMain.smoothScrollTo(mid.toInt(), 0)
         }
     }
 
@@ -250,6 +259,7 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
             SharedPreferenceManager().deleteData("email")
             SharedPreferenceManager().deleteData("password")
             SharedPreferenceManager().deleteData("accessToken")
+            ApplicationClass.categoryLists.clear()
 
             finish()
             startActivity(Intent(this@MainActivity, LoginActivity::class.java))
@@ -446,9 +456,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         // 식물 배치 공간에 클릭 이벤트 설정
         binding.mainLayout.plantPlacing1.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place1") != -1) {
-                binding.mainLayout.plantPlacing1.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing1.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place1")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing1.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing1.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place1")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing1, "1")
@@ -458,9 +477,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing2.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place2") != -1) {
-                binding.mainLayout.plantPlacing2.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing2.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place2")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing2.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing2.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place2")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing2, "2")
@@ -470,9 +498,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing3.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place3") != -1) {
-                binding.mainLayout.plantPlacing3.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing3.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place3")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing3.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing3.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place3")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing3, "3")
@@ -482,9 +519,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing4.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place4") != -1) {
-                binding.mainLayout.plantPlacing4.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing4.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place4")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing4.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing4.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place4")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing4, "4")
@@ -494,9 +540,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing5.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place5") != -1) {
-                binding.mainLayout.plantPlacing5.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing5.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place5")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing5.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing5.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place5")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing5, "5")
@@ -506,9 +561,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing6.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place6") != -1) {
-                binding.mainLayout.plantPlacing6.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing6.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place6")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing6.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing6.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place6")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing6, "6")
@@ -518,9 +582,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing7.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place7") != -1) {
-                binding.mainLayout.plantPlacing7.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing7.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place7")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing7.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing7.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place7")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing7, "7")
@@ -530,9 +603,18 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
         binding.mainLayout.plantPlacing8.setOnClickListener {
             if (SharedPreferenceManager().getInt(userEmail+"Place8") != -1) {
-                binding.mainLayout.plantPlacing8.setImageResource(R.drawable.plant_empty)
-                binding.mainLayout.plantPlacing8.startAnimation(animation)
-                SharedPreferenceManager().deleteData(userEmail+"Place8")
+                if (SharedPreferenceManager().getString("plantPlacing").equals("ON")) {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setTitle("식물 배치 취소")
+                        .setMessage("식물을 화단에서 제외하겠습니까?")
+                        .setPositiveButton("확인") { dialog, id ->
+                            binding.mainLayout.plantPlacing8.setImageResource(R.drawable.plant_empty)
+                            binding.mainLayout.plantPlacing8.startAnimation(animation)
+                            SharedPreferenceManager().deleteData(userEmail+"Place8")
+                        }
+                        .setNegativeButton("취소") { dialog, id -> }
+                    builder.show()
+                }
             }
             else {
                 val plantPlacingDialog = PlantPlacingDialog(this@MainActivity, binding.mainLayout.plantPlacing8, "8")
